+++ src/languages/r.js
@@ -18,13 +18,27 @@ export default function(hljs) {
   // handled in a separate mode. See `test/markup/r/names.txt` for examples.
   // FIXME: Support Unicode identifiers.
   const IDENT_RE = /(?:(?:[a-zA-Z]|\.[._a-zA-Z])[._a-zA-Z0-9]*)|\.(?!\d)/;
-  const SIMPLE_IDENT = /[a-zA-Z][a-zA-Z_0-9]*/;
+  const NUMBER_TYPES_RE = regex.either(
+    // Special case: only hexadecimal binary powers can contain fractions
+    /0[xX][0-9a-fA-F]+\.[0-9a-fA-F]*[pP][+-]?\d+i?/,
+    // Hexadecimal numbers without fraction and optional binary power
+    /0[xX][0-9a-fA-F]+(?:[pP][+-]?\d+)?[Li]?/,
+    // Decimal numbers
+    /(?:\d+(?:\.\d*)?|\.\d+)(?:[eE][+-]?\d+)?[Li]?/
+  );
+  const OPERATORS_RE = /[=!<>:]=|\|\||&&|:::?|<-|<<-|->>|->|\|>|[-+*\/?!$&|:<=>@^~]|\*\*/;
+  const PUNCTUATION_RE = regex.either(
+    /[()]/,
+    /[{}]/,
+    /\[\[/,
+    /[[\]]/,
+    /\\/,
+    /,/
+  );
 
   return {
     name: 'R',
 
-    // only in Haskell, not R
-    illegal: /->/,
     keywords: {
       $pattern: IDENT_RE,
       keyword:
@@ -56,6 +70,7 @@ export default function(hljs) {
         'standardGeneric substitute sum switch tan tanh tanpi tracemem ' +
         'trigamma trunc unclass untracemem UseMethod xtfrm',
     },
+
     contains: [
       // Roxygen comments
       hljs.COMMENT(
@@ -69,7 +84,7 @@ export default function(hljs) {
               // preventing highlighting. This code is example R code, so nested
               // doctags shouldn’t be treated as such. See
               // `test/markup/r/roxygen.txt` for an example.
-              className: 'doctag',
+              scope: 'doctag',
               begin: '@examples',
               starts: {
                 contains: [
@@ -89,12 +104,12 @@ export default function(hljs) {
             {
               // Handle `@param` to highlight the parameter name following
               // after.
-              className: 'doctag',
+              scope: 'doctag',
               begin: '@param',
               end: /$/,
               contains: [
                 {
-                  className: 'variable',
+                  scope: 'variable',
                   variants: [
                     { begin: IDENT_RE },
                     { begin: /`(?:\\.|[^`\\])+`/ }
@@ -104,11 +119,11 @@ export default function(hljs) {
               ]
             },
             {
-              className: 'doctag',
+              scope: 'doctag',
               begin: /@[a-zA-Z]+/
             },
             {
-              className: 'meta-keyword',
+              scope: 'keyword',
               begin: /\\[a-zA-Z]+/,
             }
           ]
@@ -118,7 +133,7 @@ export default function(hljs) {
       hljs.HASH_COMMENT_MODE,
 
       {
-        className: 'string',
+        scope: 'string',
         contains: [hljs.BACKSLASH_ESCAPE],
         variants: [
           hljs.END_SAME_AS_BEGIN({ begin: /[rR]"(-*)\(/, end: /\)(-*)"/ }),
@@ -131,48 +146,88 @@ export default function(hljs) {
           {begin: "'", end: "'", relevance: 0}
         ],
       },
+
+      // Matching numbers immediately following punctuation and operators is
+      // tricky since we need to look at the character ahead of a number to
+      // ensure the number is not part of an identifier, and we cannot use
+      // negative look-behind assertions. So instead we explicitly handle all
+      // possible combinations of (operator|punctuation), number.
+      // TODO: replace with negative look-behind when available
+      // { begin: /(?<![a-zA-Z0-9._])0[xX][0-9a-fA-F]+\.[0-9a-fA-F]*[pP][+-]?\d+i?/ },
+      // { begin: /(?<![a-zA-Z0-9._])0[xX][0-9a-fA-F]+([pP][+-]?\d+)?[Li]?/ },
+      // { begin: /(?<![a-zA-Z0-9._])(\d+(\.\d*)?|\.\d+)([eE][+-]?\d+)?[Li]?/ }
       {
         relevance: 0,
-        className: {
-          2: "number"
-        },
         variants: [
-          // TODO: replace with negative look-behind when available
-          // { begin: /(?<![a-zA-Z0-9._])0[xX][0-9a-fA-F]+\.[0-9a-fA-F]*[pP][+-]?\d+i?/ },
-          // { begin: /(?<![a-zA-Z0-9._])0[xX][0-9a-fA-F]+([pP][+-]?\d+)?[Li]?/ },
-          // { begin: /(?<![a-zA-Z0-9._])(\d+(\.\d*)?|\.\d+)([eE][+-]?\d+)?[Li]?/ }
           {
-            // Special case: only hexadecimal binary powers can contain fractions.
+            scope: {
+              1: 'operator',
+              2: 'number'
+            },
             match: [
-              /[^a-zA-Z0-9._]/, // not part of an identifier
-              /0[xX][0-9a-fA-F]+\.[0-9a-fA-F]*[pP][+-]?\d+i?/
+              OPERATORS_RE,
+              NUMBER_TYPES_RE
             ]
           },
           {
+            scope: {
+              1: 'operator',
+              2: 'number'
+            },
             match: [
-              /[^a-zA-Z0-9._]/, // not part of an identifier
-              /0[xX][0-9a-fA-F]+(?:[pP][+-]?\d+)?[Li]?/
+              /%[^%]*%/,
+              NUMBER_TYPES_RE
             ]
           },
           {
+            scope: {
+              1: 'punctuation',
+              2: 'number'
+            },
             match: [
-              /[^a-zA-Z0-9._]/, // not part of an identifier
-              /(?:\d+(?:\.\d*)?|\.\d+)(?:[eE][+-]?\d+)?[Li]?/
+              PUNCTUATION_RE,
+              NUMBER_TYPES_RE
+            ]
+          },
+          {
+            scope: { 2: 'number' },
+            match: [
+              /[^a-zA-Z0-9._]|^/, // not part of an identifier, or start of document
+              NUMBER_TYPES_RE
             ]
           }
         ]
       },
+
+      // Operators/punctuation when they're not directly followed by numbers
       {
-        // infix operator
-        begin: '%',
-        end: '%'
+        // Relevance boost for the most common assignment form.
+        scope: { 3: 'operator' },
+        match: [
+          IDENT_RE,
+          /\s+/,
+          /<-/,
+          /\s+/
+        ]
       },
-      // relevance boost for assignment
+
       {
-        begin: regex.concat(SIMPLE_IDENT, "\\s+<-\\s+")
+        scope: 'operator',
+        relevance: 0,
+        variants: [
+          { match: OPERATORS_RE },
+          { match: /%[^%]*%/ }
+        ]
       },
+
       {
-        // escaped identifier
+        scope: 'punctuation',
+        relevance: 0,
+        match: PUNCTUATION_RE
+      },
+
+      {
+        // Escaped identifier
         begin: '`',
         end: '`',
         contains: [